split apart tests and added new multistart gradient test
